Update to zls 0.9.0
Compatible with zig 0.9.0

Instead of manually editing the config to place
build_runner.zig in a custom location, this simply places
it right next to the executable in /usr/bin

This avoids having to manually modify the config.

Also, this adds an install script which prompts the user
to run `zls config` to setup the initial configuration.

I highly recomend running it :)