Include PYC files for all opt levels
and ensure that no references to the source directory are included with
the -d option. Otherwise pacman would warn about these with

==> WARNING: Package contains reference to $srcdir
usr/lib/pacautomation/__pycache__/pacautomation.cpython-310.pyc

Python generates PYC files when importing a module which is unlikely
to happen for pacautomation. pacautomation.py is only ever run directly
as a script in this package, currently. However if for whatever reason a
user or this package were to import it at runtime with root privileges,
we are already prepared. Including all optimization levels ensures that
the __pycache__ dir and  all contained PYC files are removed when
pacautiomation is removed.

[1] https://bugs.archlinux.org/task/61329
[2] https://github.com/archlinux/svntogit-packages/blob/packages/hplip/trunk/PKGBUILD