Change Dependencies Of Systemd Service Unit
Using `Upholds` here because the anki-sync-service needs nginx for
unchunking purposes as specified in the github project.

I only checked the manual to find this option now so I'm only adding it now.

----

Other ones considered:
`Wants` : This was the default choice but there was an even better
option listed further down the man page.

`Requires` : This server doesn't technically require nginx in order to
fully start itself.  It just doesn't fully function without it.

`Requisite` : For the reason above it makes sense then to try and at
least start nginx when starting this service which is what Want does.

`BindsTo` : A superset of `Requires`.  Sounds like a nice idea at first
glance but, I think it might help in diagnosing problems if only one or the other
went down rather than both.