upgpkg: python-tpm2-pytss 1.0.0-1: upstream release
This is a complete rewrite of tpm2-pytss from SWIG to CFFI. Sync the
dependencies and the build procedure with the python-tpm2-pytss-git AUR
package, which gradually introduced the necessary changes over time. Also
switch the sources from PyPI to Github in order to be able to verify their
authenticity using the PGP-signed tag.