sway-im: bump to 1.7-1, add input method v2 popups patch
Update to the latest stable release and add the input method v2 popups patch.
The popups patch is the last bit that is missing for full IME frameworks support in Wayland.
Currently, it doesn't look like wlroots changes will be needed, so I
think it's better to stay on stable and just port back the patch, if needed, so _pkgver is gone.
A minor issue that was fixed is the installation folder of the license. It should match the package name.
This confirmed to be working with anthywl, though I'm not sure how well, as I don't really know how to
type in Japanese. It doesn't seem to work with recent dev build of Fcitx5.