thedarkmod-bin: switch to a mirror which still has the old file structure
Upstream has radically changed their mirror structure, apparently in
preparation for the upcoming 2.10 release: there is no longer a single
directory containing all the necessary binaries and assets for a given version
of The Dark Mod. Instead, every release directory only contains the files that
have changed since the last release. In order to get all necessary files, one
needs to start at the manifest file

https://update.thedarkmod.com/zipsync/tdm_installer.ini

From there, find the current release and walk back through the older releases
to get all necessary files.

The upstream installer
(https://update.thedarkmod.com/zipsync/tdm_installer.linux64.zip) does this by
parsing custom "zipsync" manifest files named "manifest.iniz" that are referred
to by the above "tdm_installer.ini". However, the installer is a GUI tool that
is unsuitable for packaging use.

It would theoretically be possible to collect all the necessary files for use
in a PKGBUILD by analysing the "tdm_installer.ini" and linked "manifest.iniz"
files. This rather tedious process needs to be repeated for each release. If
someone wants to look into this, a good starting point might be the mirror

http://mirror.mstrmnds.me/releases/zipsync/

Unlike most other mirrors, it has directory listings enabled, which makes it
easier to find the necessary URLs without diving into the source code of
tdm_installer.

Due to the changes detailed above I am not going to update this package any
further: the structure of the game was never well-suited for system-wide
packaging in the first place since it requires a user-writeable game directory,
and these changes make it even harder to keep updating it.

However as a last effort to keep this package alive for a little bit longer, I
have changed the download mirror to one that still serves the old file
structure. I am unsure of how long this mirror will still be available, and I
am pretty sure the next release will not be available from there.

As an alternative to this package, I recommend directly using the upstream
installer instead, which will be kept in sync with upstream development.