Bump version to 2, fix freezing issue
Rationale behind bumping the version to 2:

Even though the goggame-1166314684.info file claims it’s version 1,
the start screen of the game says “Demo v2”. Similarly, the library
entry on gog.com also says “V2”.
What the game itself says should have priority over what the metadata
says, so change the version to 2.

About the freezing issue:

After completing the first match of the tutorial, the game would hang.
The freeze happens at the precise moment where the map is supposed to
appear.
A couple of strace runs reveal that at the point where the map is
supposed to appear, the game is trying to save its progress for the
first time. It tries to do so directly in the application directory,
which in turn causes the game to freeze because /opt/inscryption is not
user-writable.

Fix this by creating our symlinks one level further down the hierarchy.