Streamline upstream package handling
The part that handles the upstream package is a little more complex
in other GOG games, such as the full version of Inscryption [1].

Use the same (more complex) handling here as for the full version.
That minimizes drift between the PKGBUILDS, even though the additional
complexity is not really needed for the demo version.

[1]: https://aur.archlinux.org/packages/gog-inscryption