Bump pkgrel for small changes
No need to update if you already have this built. This is so that
the AUR database will pick up a new `provides` string.

I've also cleaned up the mix of tabs and spaces previously used
in the file, and fixed a broken test - these are now enabled by
default.