Create lei-git package
This PKGBUILD is based on the public-inbox-git PKGBUILD, but only has
the dependencies required for running lei. The public-inbox binaries and
manpages are built as well since there's no way around it currently in
public-inbox's Makefile, so they're deleted after the fact.

This package conflicts with public-inbox and public-inbox-git since the
perl modules from public-inbox are kept around for lei.