fix license files (changes every download?)
It appears that there are elements of the pages
https://discord.com/terms and https://discord.com/licenses that are not
completely static. This causes issues with file integrity checks. Rather
than include them in the sources array, this change downloads them with
curl during prepare(). As an added benefit, the .html files are
transformed with html2text (python) to be human-readable.