Update to v1.5.6
Maybe due to the .NET 5 update the gog package is somewhat broken,
complaining it can't find the .a versions of some libraries as required
by the .deps.json file. Provide a version of that file with those
dependency references removed as it seems to work just fine with the
dynamically linked versions.