Redirect to forked version.
Adopted this orphan.
Update PKGBUILD to change karan --> jeremymreed in URLS.
Update sha256sums.
Update build() hook, use the included tool.sh script to install deps and
build the software.
Update joe.install joe --> joe-gitignore, since the PKGBUILD renames the
binary to joe-gitignore.  This is likely because there is another
package called 'joe' in the AUR, and it looks like an editor or some
kind.  Name was likely changed to avoid collisions.