Add dependency on xorg-fonts-encodings
AUR user produnis [1] reported a list of packages that helped to get bb
running without errors; out of those, I only have xorg-fonts-misc and
xorg-fonts-encodings installed, and since xorg-fonts-misc is already in
the depends, I’m assuming xorg-fonts-encodings is the one missing
dependency. Let’s add it and see if it helps people.

[1]: https://aur.archlinux.org/account/produnis