Revamp versioning scheme
The build number is no longer the $pkgrel. From now on, the build
number will be incorporated into $pkgver as it appears in the download
link on each GitHub release (with underscores in place of dashes, since
dashes are not allowed). This allows packagers more flexibility in
tweaking the PKGBUILD.