Relax `gnome-keyring` dependency [MAJOR CHANGE]
There's no real reason for this package to have a hard dependency on
`gnome-keyring`, other than it being specified directly in the Debian
package. If this doesn't work for you, holler on the AUR package page
and/or send a GitHub issue while keeping me in the loop. (Please make
sure to specify your `org.freedesktop.secrets` provider when doing so.)