install zsh completion with package
Due to a file conflict with the bash-completion package, only zsh
completion is installed automatically with this package. Upstream
bash-completion has deprecated their completions in favor of the
generated completions, but a new release has not been cut with this
fix.

bash users seeking command completion will need to add completion
directly to their .bashrc files, which will override the completion
provided by the bash-completion package.