Migrate tests to Python 3
and use python-nose (instead of removed python2-nose)

The patch may not be ideal. It converts bytes to str so that most code
can be still used. And in some necessary cases it converts back to
bytes.
Using bytes everywhere would probably be better, but it would require
much more changes (and I am not a Python person).