Replace polkit with seatd dependency
Since seatd has become a requirement for sway's logind support, we can leverage
this package to also obtain elevated system privileges instead of polkit.

Note that using seatd instead of polkit means that a user must be a part of the
"seat" group: Otherwise, the user will not be able to achieve the appropriate
privileged actions.

Also remove a few unnecessary extra packages from the optdepends to give polkit
a bit more visibility.