Add python-requests as dependency
Previously missing, this dependency would not show up even in a
clean chroot build, as it only shows up on running the script,
and only if python-requests is missing. Installing a large set
of packages by installing python-pip would also satisfy this
dependency, which would hide that it needs to be explicitly
declared.