Fix pkgver so it uses the correct version base
In the gwenhywfar repository, the Git tag that we want is usually not
reachable from the current HEAD.
That causes `git describe` to report a version base that is too old.

Even worse, downstream packages that depend on gwenhywfar, such as
aqbanking-git, are often failing at configuration time because they
look at gwenhywfar’s pkgver, too.

To work around that issue, use the following alternative sources to
calculate a more precise pkgver:

- Extract the version string from upstream’s configure.ac instead
  of `git describe`.

- Use `git rev-parse` to count commits that have been made since
  the tag. (This assumes that an actual Git tag exists that matches
  the version string from configure.ac, which is usually the case.)

- Use `git describe` only for the abbreviated commit hash that forms
  the last pkgver segment.