update unbundling patch for libtiff 4.4.0
libtiff 4.4.0 no longer has _TIFFDataSize, so the patch needs to be updated
accordingly:
https://gitlab.com/libtiff/libtiff/-/commit/11f3f279608ae9e68f014717393197f430f9be58

Also explicitly depend on libtiff to make this direct dependency more obvious.

Compared to the patch in freeimage 3.18.0-13 we do not need the external
declaration any more since we are now linking against the correct version of
libtiff that has the new TIFFFieldSetGetSize.