Clear PKGBUILD: A user was re-pushing this despite being obsolete
This package has been deleted earlier from AUR,
because its dependency firefox-beta was not updated since
October 2021, and has been deleted in March 2022.

A user was then re-pushing updates to this, despite the
aforementioned firefox-beta did not exist anymore.

Even earlier this package was not installable, because it hardcoded
a minimum version dependency on firefox-beta, whereas that package
was not updated so the requirement of this package could not be
satisfied.