Remove garbage from html; use var not patch
The default build of the docsite contains a lot of stuff we don't want
for a local build, like ads, marketing links, tracking JS, etc. We
remove these with cli args to Sphinx.

Also, it's not necessary to patch the makefile to ensure that we don't
use more than 1 cpu during the build; the argument CPU=1 will be
propagated to subsequent make invocations.