Initial commit. YARsync v0.1.1.
Note: this package is different from the previous yarsync package on AUR!
See my comment at https://aur.archlinux.org/packages/yarsync#comment-872103

"""From 2018 (when it was created) this was a package by Alxe (https://github.com/A-lxe/yarsync). It was not very popular, so I hope there will be no issues for other people.

I created a different package with exactly same name, so don't mess these two ones! Thanks to the author for allowing me to take this AUR name!
"""