Enable systemd logging target
As recommended by one of the upstream maintainers, this commit enables
systemd-journald logging to match the builds on other platforms. If
upstream were to make linking the systemd library runtime configurable
this could be moved to optdepends, but I think currently needs to be a
full dependency if the switch is enabled at compile time.