all: set the DESTDIR variable when running build.sh
The goal is to allow us to inspect the installed files and their
structure.

The DESTDIR directory contains installed files, similar to prefix.
If its not set, the directory will be created under /tmp with random
name.

This also fix build error due to META-INF directory exist on the second
or next build.