Get `Unciv.desktop` and `Unciv.sh` from release
Previously the `Unciv.desktop` and `Unciv.sh` files were fixed in this
repository. They are now downloaded directly from the GitHub release so
they are always up-to-date.

Furthermore, changed the name of the icon to `unciv.png` instead of
`Unciv.png`, hopefully making gnome desktops able to find it and link it
to the correct desktop application.