Switch to --build/--install via CMake
Make it look more like the official package. But use Make instead of
Ninja, since Make is already installed for users of the AUR (because
of the implicit base-devel requirement). No need to require another
build dependency, especially when the project is this small and build
times are virtually indistinguishable between the two build systems.