Drop lgogdownloader build dependency
Having the lgogdownloader as a hard build dependency makes building
difficult for users who prefer to download the source files manually
from GOG, instead of using a downloader. Solve this by dropping the
dependency, but keeping the DLAGENTS for lgogdownloader in the
PKGBUILD. This way nothing changes with the build experience for
lgogdownloader users, but now there is no extra hassle for manual
downloaders.