sources: add patch for python3.10 ssize macro error
This issue was caused by a change in the way Python 3.10 handled a macro
definition.

It was fixed in a pull request in the upstream ceph repo [0] for an
unrelated build issue.

This commit adds a patch file porting that fix and updating the PKGBUILD
to include it during builds.

[0]: https://github.com/ceph/ceph/pull/44112