repo/justfile: add mkchroot, chroot recipes
for creating the base chroot that will be used in package builds.
mkchroot *only* ensures that the base chroot exists, but does not sync
it with any package updates. chroot does both, and in the future it is
expected that the standard flow for first runs of this repo will be:

  just chroot build