pkg: 15.2.14-7
- Created the base PKGBUILD
- Regenerated SRCINFO

This repo is a thin shim around the real package, ceph. It's primary
purpose is to provide prebuilt ceph artifacts, rather than building from
source.

This repo's "build" process should never be more complicated than:

1. Downloading the built ceph artifacts
2. Installing them

Any changes / fixes / discussion on the actual build of ceph should take
place in:

 https://aur.archlinux.org/packages/ceph

This is the initial release, however for sanity we're keeping the pkgrel
in sync with the upstream aur package (ceph).