Replace `flit` invocation with `build`
The `flit` command performs some up-front validation before it actually
builds the package. This is intended for upstream project maintainers
so they have a chance to learn about issues before they proceed to cut
a release.

Once a package is released, people who consume the package (which
includes package repository maintainers) are assumed to be mainly
interested in getting the already-released package to build. Therefore,
they may simply want to run the build backend without caring about
up-front validation. That’s what the `build` command does: invoke the
suitable build backend, which in the case of `flit`, doesn’t run the
validation steps.

See discussion in [1].

[1]: https://github.com/pypa/flit/issues/580

CC: Thomas Kluyver <thomas@kluyver.me.uk>
CC: William Woodruff <william@yossarian.net>