Remove patch enabling spice-common to detect pyparsing
In spice-common a bug preventing it from detecting recent versions of
pyparsing [1] was fixed in a7b5474b [2].
The problem was also affecting spice [3], as it is including
spice-common as a Git submodule.
When package spice-git was introduced to the AUR, upstream spice was
using a spice-common commit before a7b5474b, so patching was needed.
Meanwhile, upstream spice is using a commit of spice-common including
the fix, so the patch has to be removed from package spice-git.

[1] https://gitlab.freedesktop.org/spice/spice-common/-/issues/5
[2] https://gitlab.freedesktop.org/spice/spice-common/-/commit/a7b5474bf808934cf0ee1107a58d5f4d97b9afbf
[3] https://gitlab.freedesktop.org/spice/spice/-/issues/70