Bump pkgrel to reinstate wrongfully deleted package
This package was deleted because "the files contained in this package
are provided by edk2-armvirt". This is false; edk2-armvirt provides the
files for aarch64, and this package provides the files for 32-bit ARM.
Therefore, I am attempting to reinstate this package by bumping its
pkgrel, as suggested by Kevin Morris.