pkgbuild: fix pkg deps, world readable dirs
There seems to be a lot of changes in the required packages since 15.x,
which have now been fixed / updated based on namcap's output

This appears to be causing issues with other things that want to read
ceph.conf and other files. There shouldn't be any security concerns
here, as all keyrings are generated 0600 by default.