always restart on SIGALRM
k5start uses SIGALRM internally, so a common hack is to send it one from
the outside (via `pkill -ALRM k5start` from NetworkManager, for example)
to force renewals. However, sometimes that occurs when k5start *doesn't*
have its SIGALRM handler set up, so the result is that the service just
crashes. (It's not a good hack.)