init: Add PKGBUILD for packaging retsnoop to AUR
Arch Linux allows users to package binaries in the AUR by defining a
PKGBUILD file. This file needs to be updated in the AUR when new retsnoop
releases are made in order for users to use new versions of retsnoop. This
initial commit adds the first PKGBUILD file for packaging the retsnoop
project, along with the generated .SRCINFO file.

Signed-off-by: David Vernet <void@manifault.com>