Remove custom permission handling
This was introuce in the first commit (3ccf7959ec0b). Since then it
always make trouble if some new executables was introuced inside
electron on which Gitkraken depends.
There are no clear statements to this inside wiki for example regarding
security concerns.
https://wiki.archlinux.org/title/PKGBUILD
https://wiki.archlinux.org/title/creating_packages

This custom permission is removed to address current issues regarding
the upgrade.

Close #207