PKGBUILD: fix generated package file
Since we're overriding $pkgname, the generated output filename currently
is `ddclient-v3.10.0.r0.g3e2cb0a-1-any.pkg.tar.zst`. This is incorrect
because we lost the `-git` suffix.

Ensure that the $pkgname is not overridden by the end of package().

This is cherry-picked from d6cf1cd (PKGBUILD: fix generated package
file, 2022-10-24) because I've made the same mistake twice.

Reported-by: jahway603 <jahway603@protonmail.com>