pkgbuild: pkgver=17.2.5
Normally I don't do point upgrades in the same patch set as major
upgrades, however the upstream made an 'oops' in their release train for
v17 that lead to a couple important regressions. See the changelog below
for an overview, and the link for the full story.

Changlog:

* c6e18266dfb mgr/telemetry: handle daemons with complex ids
* 6a1a427fc5f ceph-volume: fix regression in activate
* feeadd7e2a8 mgr/rook: fix error when trying to get the list of nfs services
* 1d8a2b46bd3 Revert "osd/PeeringState: fix missed `recheck_readable` from laggy"
* c67f74c3c76 Revert "osd/PeeringState: proc_lease_ack break once found from OSD"

Link:

  https://ceph.io/en/news/blog/2022/v17-2-5-quincy-released

Upstream-Ref: https://github.com/ceph/ceph/pull/48519