pkgbuild: fix-iterator-depreciations.patch
This commit patches the few instances of std::iterator usage in the
upstream.

There's no real benefit here, besides reducing the line noise during
builds... and yet this is one of the change sets I've liked making the
most.

Followed the fluentcpp guide for understanding the fix:

  https://www.fluentcpp.com/2018/05/08/std-iterator-deprecated/

and used this ripgrep to find the definition sites:

  `rg ':\s*(public)?\s*(std::)?iterator\s*<' .`