pkgbuild: add ceph.sudoers to package_ceph()
This file is inspired by the upstream's 'sudoers.d/ceph-smartctl' file,
however, unlike their one; isn't a security threat if anyone nasty ever
elevates to the 'ceph' user.

We use the latest and greatest in sudo technology: regexes. Sudo is only
4 decades late to the party, and we welcome them in with open arms,
though unfortunately all the pizza has been eaten, along with most of
the soda drinks, and most of the party's participants are tired, like
this metaphor.