Fix version to consider tags
Unfortunately this doesn't update cleanly because the previous version
did not follow VCS packaging guidelines and omitted the r prefix to
designate the entire version as revision number.  Hence it created a
ridicuously high version far ahead of any kind of tag.

Regardless we should fix this to get this right for the future.