Add bash-completion as a depends
Originally graysky claimed that cmake was required to build this
package... yeah, well that is what it could seem like, but really it was
becasue meson wanted bash-completion to be optionally present -- so it
searched for bash-completion via first pkg-config and then cmake to then
fail and accept it. Installing bash-completion on the system fixes the
issues.