Exclude base/system_info001 until next release
The base/system_info001 test has a bug that was not discovered until
recently, due to the systems we (the Idris community) have access to not
showing it. It has been fixed as of idris-lang/idris2#2754, but that
won't land in this package until the next official release.

Also removed git as a makedepends. It was likely left over from the
idris2-git package; the official release doesn't need git to build.