mingw-w64-waffle 1.7.0-2: build with meson
The meson build depends on cmake to generate the cmake files.
But it must not use the mingw-w64-cmake wrapper, since that fails to
print the basics needed. Aka --system-information with the wrapper
outputs nothing.

In addition, we need to disable LTO since it completely stuffs the final
binaries. Similarly - stripping is left to the packaging stage.

There are some caveats, comparing the cmake build:
 - the meson cross-file set ssp, so one needs libssp-0.dll
 - upstream uses __thread, so one needs libwinpthread-1.dll
 - upstream sets the prefix, so locally we rename {lib,}waffle-1.dll

The latter two be fixed with a later upstream release.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>