Lessen sandbox restrictions
The new DNS feature requires modifying system configuration files
(through systemd-resolved), which doesn't play nice with our attempt to
run netbird daemon as a regular user.

Let's also use this chance to get rid of the old unit alias
(wiretrustee@.service). It was supposed to exist for a month or two
after the project was renamed.