add patch for enabling legacy ciphers
See https://github.com/mtrojnar/osslsigncode/issues/178

The short is that certificates stored in PKCS12 containers
use RC2-40-CBC, which is disabled by default in OpenSSL 3.0.
There is a PR that enables the legacy ciphers for osslsigncode:

https://github.com/mtrojnar/osslsigncode/pull/194

The first commit adding the functionality is applied against
release 2.5 and works fine. The second commit which only
adds further tests does not apply against 2.5, though, and
is therefore not applied by the PKGBUILD.