strip package contents to only init scripts
It does not make a lot of sense to re-package the whole greetd
just to provide init scripts.
I originally created the package to mirror greetd-artix-runit,
which does the whole repackaging thing.

AFAICT, Artix nowadays use the approach of main package with the daemon,
and various -openrc/-runit/-… packages shipping only init scripts
and depending on the main daemon.