Convert into split package with TTF and OTF variants
As per user request, split the package in TTF and OTF variants, so it is
possible to choose which font format to install. While at it, make yet
one more split package containing the Fontconfig configuration file,
with both packages depending on it. This is a first step towards adding
more format variants.