Install binary to /usr/bin and remove wrapper.
This removes the old wrapper script. New versions of supermodel offer
propper multi-user support on Linux out of the box.

This should be compatible with most configurations. It supports
pre-existing configurations in ~/.supermodel and ~/.config/supermodel.

Scrips which relied on the old binary location at /usr/share may need
to be modified to use the new path.

I also old patches that aren't being used anymore. But we can retrieve
them from version control if necessary.