Change name of download file to pass checksum
Some users are reporting a checksum failure because upstream moved the
repo name and thus the embedded dir in the github tar file has changed
even though the file has the same name and version. Thus if you had an
old copy of that file then it would not be downloaded and would fail the
checksum check. Changing the file name forces it to be re-downloaded.