upgpkg: 1.4.2 + deps reduction
qt5-webengine is no longer used in 1.4.2. This is a good opportunity to
slim down the depedencies, as without qt5-webengine, only qt5-multimedia
is worth depending on. The other qt5-* deps only depend (directly or
indirectly) on qt5-base, so requiring qt5-multimedia is enough to pull
the non-qt5 deps we actually need.

Also stop depending on desktop-file-utils (like xnviewmp-system-libs),
as it is not specifically needed by this package.