fix: turn off desktop file parameters substitution
The $HOME variable was expanded during desktop file
creation because parameter substitution was left in
the heredoc syntax. So the software config path was
the one of the user who installed the package, even
if another user was running the executable.
This correction therefore allows correct execution
of the software on a multi-user workstation.