Pull request #2 - Migration of data when updating OneDev package
As brought up in issue #1 onedev data was not being migrated when the package was updated, this now works provided you overwrite the changed files from the update.sh script.

The package now provides two packages:

onedev -> the main package containing the onedev users, and systemd scripts

onedev-new -> package containing only the new content, the update.sh script should be called on /opt/onedev, and then you can remove onedev-new and install onedev

This is a working workaround for the current JSW issues, this is very raw at the moment, but can be ironed out along the way, at least it is better than nothing.

As a sidenote this can not be merged as the systemd service file no longer functions correctly!