feat: New AUR package of "commit" a command line aplication written in rust
I have changed the of name of the package because already exists one in the AUR named commit, some others packages do this and I think it is fine. I have read all the rules to publish packages to the AUR and I think this program can be useful for a majority of the people.