Add missing dependency on wheels
There is a slightly odd dependency issue with `python-build`. The issue is
`python-build` produces either source distribution or wheel, so the wheel
support is built-in. Moreover, `python-installer` does not depends on
`python-wheel` as well because it directly implements "wheel protocol".
However, `python-build` uses system wheel package or install it into isolated
environment as if there could be different wheel implementations. Not
clear.