pkgbuild: add rdkafka.mpatch
This is a fix to confluentinc/librdkafka#4155, which blocks our ability
to build this package.

The plan is to apply the patch to the system file, output into a local
directory and instruct pkgconfig to search that directory.

References: https://github.com/bazaah/aur-ceph/issues/11
References: https://github.com/confluentinc/librdkafka/issues/4155