new release 0.4.7
WARNING: needs a clean repo to not show '-dirty' in `version` subcmd

new version/tag behaviour, currently problematic because real version is
not shown in `version` subcmd anymore.
Reason might be because the real versoin is implicitly taken from a git
tag by the go toolchain by default. But as it's not build from the
original git repo the git tag which carries the version information is
missing.