Fix gradle not able to find javac
This issues seems to happen regardless of the gradle version,
except maybe the one that can be installed with the included wrapper.
This seems to be caused by the setting
`options.forkOptions.executable = 'javac'`
in build.gradle.
This commit removes this option,
as the comment suggests it is only done to fix warnings.

Maybe this should be fixed upstream, either on SheepIt client or gradle,
but not sure where, why or how.