Use distributed flac, sdl, and smpeg
libFLAC.so.8 is no longer in lib32-flac; it is available in
lib32-flac1.3, but since the 64 bit version is also only available on
AUR using the distributed version seems less brittle. Using the system's
libSDL breaks video. libsmpeg-0.4.so.0 is no longer in smpeg, and
there's not already a package on AUR.