molly-brown: add config to the repo
The example just slightly updated in regard of the log files.
Default the access.log to /var/log/molly-brown, so that adding the tmpfiles integration and adding the folder makes sense.
Although not enabled, change the location for the error.log to its actual default: if empty, then it prints to stderr, which then appears in the journalctl