create version from toml
work around a deficiency in the git repository tagging strategy
by retrieving the version out of the cargo toml. strip "unstable"
as the package name is gossip-git which by itself means it is
tip of master. the increasing number is from the last correct
tag, the hash to make the version unique.