Temporarily fix source package link
I have pushed new tags to the
github.com/ungoogled-software/ungoogled-chromium-archlinux repository
without properly bumping the pkgrel in PKGBUILD, causing a wrong release
url to be used.

This patch bumps the pkgrel to 3 (matching the upstream tag) and
hardcodes to download the -1 package (as the release package produced by
CI also depends on the PKGBUILD file for its revision).

The next release will force override any temporary changes made to the
PKBGUILD file, hopefully putting the release process on track again.

Sorry for the inconvenience!