Added ByteOnPanel
This is an old tool that was originally available as a mercurial
repository in bitbucket. After Atlassian decided to nuke all mercurial
repositories, Pavel-Sayekat mirrored it on GitHub. whorfin later forked
it to clean up a few unneeded files and to add support for bit/s.

If anyone is interested, please make the bit/byte change a command-line
argument, and submit the patch upstream.

https://bitbucket.org/mozbugbox/byteonpanel/
https://github.com/Pavel-Sayekat/ByteOnPanel
https://github.com/whorfin/ByteOnPanel

This tool is useful for minimalistic desktop environments like
LXDE/LXQT or openbox. People looking for more monitors should try
multiload-ng.