all: add some patches to work around build issues
Add some interim patches to get this building with the latest version of
cargo/rust. This is messy and will be dropped as soon as the
corresponding upstream PR is accepted [1].

Why is the patch bumping the uhid-virt version? Because the version
published to crates.io doesn't match the upstream package version,
despite the actual contents of the crate being the same. Feel free to
verify this [2].

[1] https://github.com/flukejones/uhid-virt/pull/4
[2] https://static.crates.io/crates/uhid-virt/uhid-virt-0.0.5.crate

Signed-off-by: StaticRocket <35777938+StaticRocket@users.noreply.github.com>