Update to 22.3.2
But remove aarch64 from the arch array, because there are no
corresponding binaries in the release (same as in the jdk17 version of
the package). Supposedly this is intentional, though I haven’t heard an
explanation yet for why there were aarch64 binaries in the previous
release. Let’s see if they return in future, I guess.