pkgbuild: package more client / utils
These are a couple useful utilities for interfacing with ceph, of which
ceph-volume is likely the most important, as is it responsible for
provisioning devices for use as OSDs.

cephadm is an honorable mention, being a 9k line python file used in
managing clusters... if you like the idea of having 9k lines of python
sitting between you and your entire storage backplane.

(I do not)

- ceph-volume
- cephfs-shell
- cephfs-top
- ceph-cephadm