pkgbuild: package misc. ceph tooling / libs
These are mostly used by the upstream development team for interacting
with various internals of the cluster and live tests.

However, the impetus for this entire patchset also lurks: java-cephfs.
A package that requires we pull in the entire java runtime, likely for a
package absolutely no one uses. Good riddance.

- ceph-tools
- ceph-test
- java-cephfs