Change update wrapper script to use fixed UID/GID
The helper script currently uses the onedev user/group name to chown
the live directory after creating/updating it. However on first
installation this script is run before the user/group have been
created, causing this step to fail and the installation to not work
out of the box.

This commit modifies the wrapper script to use the fixed UID and GID
specified in the sysusers configuration which fixes the above problem.