PKGBUILD: Let pacman manage sources allowing builds to be offline
Currently, the buildgcc script used to build the toolchain requires an
internet connection during the build step to download the source
tarballs. Building the package reproducible is not possible this way.

Thus, let pacman manage (and download) the source tarballs and create
symlinks to the directory where they are expected. This allows the build
step to be offline improving the reproducibility.

To keep things modular, create separate arrays for GCC and Clang,
containing sources and hashsums, and add the contents to the specified
arrays later if their components are enabled. This way, the sources are
not downloaded when their components are not enabled.

Signed-off-by: Felix Singer <felixsinger@posteo.net>