update to latest version, depend on older python
The v5.15.3-lts-lgpl tag isn't on any branch, so use the tag directly.
Also, set the pkgver to a constant so it's clearer what version is
installed. We can't use a simple `git describe` because it reports
5.11.2 for some reason.

Depend on an older version of python, since the latest version is not
compatible with this version of pyside2.